Move list functionality into ddevapp
