+++ cmd/ddev/cmd/list.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"time"
-
 	"github.com/drud/ddev/pkg/ddevapp"
-	"github.com/drud/ddev/pkg/output"
-	"github.com/drud/ddev/pkg/util"
 	"github.com/spf13/cobra"
 )
 
@@ -27,37 +23,7 @@ var ListCmd = &cobra.Command{
 ddev list --active-only
 ddev list -A`,
 	Run: func(cmd *cobra.Command, args []string) {
-		runTime := util.TimeTrack(time.Now(), "ddev list")
-		defer runTime()
-
-		for {
-			apps, err := ddevapp.GetProjects(activeOnly)
-			if err != nil {
-				util.Failed("failed getting GetProjects: %v", err)
-			}
-			appDescs := make([]map[string]interface{}, 0)
-
-			if len(apps) < 1 {
-				output.UserOut.WithField("raw", appDescs).Println("No ddev projects were found.")
-			} else {
-				table := ddevapp.CreateAppTable()
-				for _, app := range apps {
-					desc, err := app.Describe(true)
-					if err != nil {
-						util.Error("Failed to describe project %s: %v", app.GetName(), err)
-					}
-					appDescs = append(appDescs, desc)
-					ddevapp.RenderAppRow(table, desc)
-				}
-				output.UserOut.WithField("raw", appDescs).Print(table.String() + "\n" + ddevapp.RenderRouterStatus())
-			}
-
-			if !continuous {
-				break
-			}
-
-			time.Sleep(time.Duration(continuousSleepTime) * time.Second)
-		}
+		ddevapp.List(activeOnly, continuous, continuousSleepTime)
 	},
 }
